Create useActivities hook
